- Minor change in type content.

+++ src/DiffEqBiological.jl
@@ -12,7 +12,7 @@ using Parameters
 @reexport using DiffEqBase, DiffEqJump
 using Compat
 
-const ExprValues = Union{Expr,Symbol,Float64,Int64}                   
+const ExprValues = Union{Expr,Symbol,Float64,Int}                   
 
 include("reaction_network.jl")
 include("maketype.jl")
